safari-534.53-branch/LayoutTests/sputnik/Conformance/07_Lexical_Conventions/7.8_Literals/7.8.4_String_Literals/S7.8.4_A7.2_T5-expected.txt

CONSOLE MESSAGE: line 77: SyntaxError: Parse error
S7.8.4_A7.2_T5

PASS Expected parsing failure

TEST COMPLETE
